# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Dokumentasi API untuk CRUD Admin WA Rotator",
        "version": "1.0"
    },
    "paths": {
...
        "/api/category-admin": {
            "get": {
                "tags": [
                    "Category Types"
                ],
                "summary": "Menampilkan daftar Lokasi",
                "description": "Mengambil daftar Lokasi dengan fitur pencarian berdasarkan name atau deskripsi",
                "operationId": "86920feeadb51567896daedea5641941",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filter Lokasi berdasarkan name atau deskripsi",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "villa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Users retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "john_doe"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "........."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/category-create": {
            "post": {
                "tags": [
                    "Category Types"
                ],
                "summary": "Create a new category type",
                "operationId": "204a0cbd11a146b4fc33bfe3654e8e15",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Type created successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/api/category-update/{id}": {
            "put": {
                "tags": [
                    "Category Types"
                ],
                "summary": "Update an existing category type",
                "operationId": "41111e666c60d1cf015593ec69e322ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Type updated successfully"
                    },
                    "404": {
                        "description": "Type not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/api/category-delete/{id}": {
            "delete": {
                "tags": [
                    "Category Types"
                ],
                "summary": "Delete a category type",
                "operationId": "2daa6786b7c15a801a10a4a1290d29f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type deleted successfully"
                    },
                    "404": {
                        "description": "Type not found"
                    }
                }
            }
        },
...
    },
...
}